Remove the second tier of filesystems
Also:
* Reorganize namespacing and naming
