+++ tools/wf/plugin-standard.jsh.fifty.ts
@@ -82,7 +82,7 @@ namespace slime.jsh.wf {
 							path: "."
 						});
 						var slime = fixtures.clone({
-							src: jsh.file.world.spi.filesystems.os.pathname(src.toString()),
+							src: jsh.file.world.filesystems.os.pathname(src.toString()),
 							commit: {
 								message: "Local modifications"
 							}
@@ -159,7 +159,7 @@ namespace slime.jsh.wf {
 							if (!p) p = {};
 							var origin = fixture();
 							var clone = fixtures.clone({
-								src: jsh.file.world.spi.filesystems.os.pathname(origin.directory.toString())
+								src: jsh.file.world.filesystems.os.pathname(origin.directory.toString())
 							});
 							clone.submodule.update({
 								init: true,
@@ -589,7 +589,7 @@ namespace slime.jsh.wf {
 				fifty.tests.manual.profile = function() {
 					var project = test.fixtures.adapt.repository(test.fixtures.hosted({ noInitialize: true }).clone);;
 					var getSlimePath = function(relative) {
-						return jsh.file.world.spi.filesystems.os.pathname(project.directory.toString()).relative("slime").relative(relative);
+						return jsh.file.world.filesystems.os.pathname(project.directory.toString()).relative("slime").relative(relative);
 					}
 					//	TODO	should profiler install Rhino?
 					$api.fp.world.now.action(
